feat(messages): convert e2eeType to integer in message creation

+++ src/routes/messages.ts
@@ -185,7 +185,7 @@ messagesRouter.post('/',
   body('senderId').isUUID(),
   body('content').notEmpty().isLength({ max: 131072 }),      // ciphertext Signal (ECDH peut être grand)
   body('senderContent').optional().isString().isLength({ max: 131072 }),
-  body('e2eeType').optional().isIn([1, 3]),                  // type Signal
+  body('e2eeType').optional().toInt().isIn([1, 3]),                  // type Signal
   body('replyToId').optional().isUUID(),
   validateRequest,
   messageController.create.bind(messageController)
